Styling extracted as a hook
