+++ react-components/src/utilities/constants.ts
@@ -1,1 +1,4 @@
-export const DEFAULT_QUERY_STALE_TIME = 1000 * 60 * 10 ; // 10 minutes
+/*!
+ * Copyright 2023 Cognite AS
+ */
+export const DEFAULT_QUERY_STALE_TIME = 1000 * 60 * 10; // 10 minutes
